coderabbit review comment fixes

+++ nexthop.go
@@ -7,38 +7,42 @@ import (
 	"strings"
 )
 
+// Nexthop group types - reference https://github.com/torvalds/linux/blob/master/include/uapi/linux/nexthop.h
 const (
-        // NEXTHOP_GRP_TYPE_MPATH is default multi-path hash threshold
-        NEXTHOP_GRP_TYPE_MPATH uint16 = iota
+	// NEXTHOP_GRP_TYPE_MPATH is default multi-path hash threshold
+	NEXTHOP_GRP_TYPE_MPATH uint16 = iota
 
-        // NEXTHOP_GRP_TYPE_RES is Resilient nexthop group
-        NEXTHOP_GRP_TYPE_RES
+	// NEXTHOP_GRP_TYPE_RES is Resilient nexthop group
+	NEXTHOP_GRP_TYPE_RES
 )
 
 // NexthopGroupMpath represents one member of a nexthtop group
 type NexthopGroupMpath struct {
-        // ID of an existing nexthop to include in the group
-        ID uint16
-        // Relative weight, 1-256. Zero is treated as 1
-        Weight uint16
+	// ID of an existing nexthop to include in the group
+	ID uint32
+	// Relative weight, 1-256. Zero is treated as 1
+	Weight uint16
 }
 
+// NexthopResGroup is  resilient nexthop group structure
 type NexthopResGroup struct {
-        Buckets uint16
-        IdleTimer uint32
-        UnbalancedTimer uint32
-        UnbalancedTime uint64
+	Buckets         uint16
+	IdleTimer       uint32
+	UnbalancedTimer uint32
+	UnbalancedTime  uint64
 }
 
+// Nexthop represent a nexthop object
 type Nexthop struct {
 	ID        uint32
 	Blackhole bool
 	OIF       uint32
 	Gateway   net.IP
 	Protocol  RouteProtocol
-	Group []NexthopGroupMpath
+	// Nexthop group members for resilient nexthop group or multipath nexthop group
+	Group     []NexthopGroupMpath
 	GroupType uint16
-	ResGroup *NexthopResGroup
+	ResGroup  *NexthopResGroup
 }
 
 func (h *Nexthop) String() string {
